feat: simplify configuration system and improve error handling
## Configuration System Improvements
- Consolidated 4+ configuration files into single src/config.ts
- Merged duplicate validation logic from validation.ts and utils/config-validator.ts
- Removed redundant config/default-config.ts directory
- Kept all configurability features including allowed_users security
- Cleaner imports and reduced maintenance overhead

## Error Handling Improvements
- Removed all process.exit() calls for GitHub Actions compliance
- Centralized error handling with proper error classification
- Use core.setFailed() consistently throughout
- Better error messages with specific error types
- Improved debugging experience

## Code Cleanup
- Removed empty utils/ directory
- Clean dist/ rebuild without accumulated old files
- Maintained all existing functionality while reducing complexity
